Add backup restoration and deletion

+++ app/Services/Servers/Backups/BackupDeletionService.php
@@ -7,6 +7,7 @@ use Convoy\Models\Backup;
 use Convoy\Repositories\Proxmox\Server\ProxmoxBackupRepository;
 use Convoy\Services\ProxmoxService;
 use Illuminate\Database\ConnectionInterface;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
 class BackupDeletionService extends ProxmoxService
 {
@@ -16,6 +17,10 @@ class BackupDeletionService extends ProxmoxService
 
     public function handle(Backup $backup, ?bool $force = false)
     {
+        if (is_null($backup->completed_at)) {
+            throw new BadRequestHttpException('This backup cannot be restored at this time: not completed.');
+        }
+
         $this->connection->transaction(function () use ($backup, $force) {
             try {
                 $this->repository->setServer($backup->server)->delete($backup);
